requirements.txt
apipkg==1.4
autopep8==1.2.4
chardet==2.3.0
click==6.6
coverage==4.2
defusedxml==0.4.1
Django==1.10
dnspython==1.14.0
docutils==0.12
execnet==1.4.1
Flask==0.11.1
Flask-API==0.6.7
Flask-API-Utils==1.0.2
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
mosquitto==1.2.3
MySQL-python==1.2.5
paho-mqtt==1.2
pep8==1.7.0
Pillow==3.3.1
py==1.4.31
pycrypto==2.6.1
Pygments==2.1.3
pyserial==3.1.1
pytest==3.0.1
pytest-cache==1.0
pytest-cov==2.3.1
pytest-pep8==1.0.6
pytest-timeout==1.0.0
python-apt==0.9.3.12
python-debian==0.1.28
python-debianbts==1.11
reportbug==6.6.3
roman==2.0.0
six==1.10.0
SOAPpy==0.12.22
var-dump==1.1
Werkzeug==0.11.10
wstools==0.4.4
flask-jsonpify=1.5.0
flask-json=0.3.0
